+++ src/pa_mic_background.hpp
@@ -80,7 +80,7 @@ signals:
     void stopRecording(const bool &recording_is_stopped, const int &wait_time = 5000);
     void updateVolume(const double &volumePctg);
     void updateWaterfall(const std::vector<GekkoFyre::Spectrograph::RawFFT> &data,
-                         const std::vector<short> &raw_audio_data,
+                         const std::vector<int> &raw_audio_data,
                          const int &hanning_window_size, const size_t &buffer_size);
 
 public slots:
